+++ app/views/screens/welcome_screen/welcome_screen.css
