+++ app/views/widgets/help.py
@@ -5,5 +5,5 @@ from textual.app import ComposeResult  # type: ignore
 from textual.screen import Screen  # type: ignore
 from textual.widgets import ListView, ListItem, Label, Footer  # type: ignore
 
-from aws_stuff_doer.login.login import list_profiles  # type: ignore
-from aws_stuff_doer.uilib.widgets import CustomHeade  # type: ignore
+from app.login import list_profiles  # type: ignore
+from app.views.widgets import CustomHeade  # type: ignore
