--- src/aws_stuff_doer/__init__.py
@@ -1,1 +1,0 @@
-"""__init__.py for asd."""
