--- src/aws_stuff_doer/login/__init__.py
@@ -1,1 +1,0 @@
-"""Login library for AWS SSO."""
