--- src/aws_stuff_doer/open/__init__.py
@@ -1,1 +1,0 @@
-"""Open library for AWS SSO."""
