--- src/aws_stuff_doer/uilib/__init__.py
@@ -1,3 +1,0 @@
-"""The __init__.py filr for the ui module."""
-
-from .app import App  # type: ignore
