--- tests/conftest.py
@@ -1,1 +1,0 @@
-
